Rename 'commit' -> 'save'
Because 'save' is more obvious name this processor, while 'commit' may
be mistakenly assumed to be related to Git.
