+++ tests/core/test_factories.py
@@ -10,7 +10,6 @@ def test_create_app_processors_discover():
 
     assert set(testapp._processors) == {
         "archive",
-        "commit",
         "commonmark",
         "feed",
         "frontmatter",
@@ -21,6 +20,7 @@ def test_create_app_processors_discover():
         "pipe",
         "prettyuri",
         "restructuredtext",
+        "save",
         "sitemap",
         "source",
         "todatetime",
